@@ -2,6 +2,7 @@
 
 use clap::crate_authors;
 use std::io;
+use std::path::PathBuf;
 use std::thread::available_parallelism;
 use std::time::SystemTime;
 
@@ -68,6 +69,9 @@ enum Commands {
         /// The name of preset to be printed
         #[clap(required_unless_present("list"), value_enum)]
         name: Option<print::Preset>,
+        /// Output the preset to a file instead of stdout
+        #[clap(short, long, conflicts_with = "list")]
+        output: Option<PathBuf>,
         /// List out all preset names
         #[clap(short, long)]
         list: bool,
@@ -85,8 +89,11 @@ enum Commands {
         /// Print the right prompt (instead of the standard left prompt)
         #[clap(long)]
         right: bool,
-        /// Print the continuation prompt (instead of the standard left prompt)
+        /// Print the prompt with the specified profile name (instead of the standard left prompt)
         #[clap(long, conflicts_with = "right")]
+        profile: Option<String>,
+        /// Print the continuation prompt (instead of the standard left prompt)
+        #[clap(long, conflicts_with = "right", conflicts_with = "profile")]
         continuation: bool,
         #[clap(flatten)]
         properties: Properties,
@@ -117,6 +124,12 @@ fn main() {
     let _ = nu_ansi_term::enable_ansi_support();
     logger::init();
     init_global_threadpool();
+
+    // Delete old log files
+    rayon::spawn(|| {
+        let log_dir = logger::get_log_dir();
+        logger::cleanup_log_files(log_dir);
+    });
 
     let args = match Cli::try_parse() {
         Ok(args) => args,
@@ -166,12 +179,14 @@ fn main() {
         Commands::Prompt {
             properties,
             right,
+            profile,
             continuation,
         } => {
-            let target = match (right, continuation) {
-                (true, _) => Target::Right,
-                (_, true) => Target::Continuation,
-                (_, _) => Target::Main,
+            let target = match (right, profile, continuation) {
+                (true, _, _) => Target::Right,
+                (_, Some(profile_name), _) => Target::Profile(profile_name),
+                (_, _, true) => Target::Continuation,
+                (_, _, _) => Target::Main,
             };
             print::prompt(properties, target)
         }
@@ -184,21 +199,21 @@ fn main() {
                 println!("Supported modules list");
                 println!("----------------------");
                 for modules in ALL_MODULES {
-                    println!("{}", modules);
+                    println!("{modules}");
                 }
             }
             if let Some(module_name) = name {
                 print::module(&module_name, properties);
             }
         }
-        Commands::Preset { name, list } => print::preset_command(name, list),
+        Commands::Preset { name, list, output } => print::preset_command(name, output, list),
         Commands::Config { name, value } => {
             if let Some(name) = name {
                 if let Some(value) = value {
                     configure::update_configuration(&name, &value)
                 }
             } else if let Err(reason) = configure::edit_configuration(None) {
-                eprintln!("Could not edit configuration: {}", reason);
+                eprintln!("Could not edit configuration: {reason}");
                 std::process::exit(1);
             }
         }
