@@ -18,7 +18,7 @@ pub struct OSConfig<'a> {
 
 impl<'a> OSConfig<'a> {
     pub fn get_symbol(&self, key: &Type) -> Option<&'a str> {
-        self.symbols.get(key).cloned()
+        self.symbols.get(key).copied()
     }
 }
 
@@ -52,6 +52,8 @@ impl<'a> Default for OSConfig<'a> {
                 Type::NetBSD => "🚩 ",
                 Type::NixOS => "❄️ ",
                 Type::OpenBSD => "🐡 ",
+                Type::OpenCloudOS => "☁️ ",
+                Type::openEuler => "🦉 ",
                 Type::openSUSE => "🦎 ",
                 Type::OracleLinux => "🦴 ",
                 Type::Pop => "🍭 ",
