@@ -1,5 +1,5 @@
 use super::{Context, Module, ModuleConfig};
-use git_repository::commit::describe::SelectRef::AllTags;
+use gix::commit::describe::SelectRef::AllTags;
 
 use crate::configs::git_commit::GitCommitConfig;
 use crate::context::Repo;
@@ -51,8 +51,8 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
 }
 
 fn git_tag(repo: &Repo, config: &GitCommitConfig) -> Option<String> {
-    // allow environment variables like GITOXIDE_OBJECT_CACHE_MEMORY and GITOXIDE_DISABLE_PACK_CACHE to speed up operation for some repos
-    let mut git_repo = repo.open().apply_environment();
+    let mut git_repo = repo.open();
+    // Increase the default object cache size to speed up operation for some repos
     git_repo.object_cache_size_if_unset(4 * 1024 * 1024);
     let head_commit = git_repo.head_commit().ok()?;
 
@@ -373,16 +373,7 @@ mod tests {
             .output()?;
 
         let git_tag = create_command("git")?
-            .args([
-                "for-each-ref",
-                "--contains",
-                "HEAD",
-                "--sort=-taggerdate",
-                "--count=1",
-                "--format",
-                "%(refname:short)",
-                "refs/tags",
-            ])
+            .args(["describe", "--tags"])
             .current_dir(repo_dir.path())
             .output()?
             .stdout;
@@ -443,16 +434,7 @@ mod tests {
             .output()?;
 
         let git_tag = create_command("git")?
-            .args([
-                "for-each-ref",
-                "--contains",
-                "HEAD",
-                "--sort=-taggerdate",
-                "--count=1",
-                "--format",
-                "%(refname:short)",
-                "refs/tags",
-            ])
+            .args(["describe", "--tags"])
             .current_dir(repo_dir.path())
             .output()?
             .stdout;
